insecure-file-upload/requirements.txt
Flask
pydantic
python_magic
Werkzeug
